feat(backend): Refactor project structure and fix Snyk API integration
This commit introduces a major refactoring of the Go backend to align with idiomatic project layout standards and corrects the Snyk Export API integration.

Key changes include:
- The monolithic main.go has been split into a modular structure with distinct packages:
  - `cmd/nis2-dash`: Main application entry point.
  - `internal/config`: Environment variable configuration.
  - `internal/snyk`: Snyk API client and models.
  - `internal/processor`: CSV data processing.
  - `internal/server`: HTTP server and request handlers.

- The Snyk Export API integration has been updated to match the official documentation:
  - Switched to organization-level endpoints for creating and polling export jobs.
  - Corrected the export request payload, including the 'type', filter names, and removed the 'destination' field.
  - Updated polling logic to use the correct job status and result retrieval endpoints.

- Corrected the module path in [go.mod](cci:7://file:///Users/iuriikogan/Development/github.com/iuriikogan-snyk/NIS2-Dash/backend/go.mod:0:0-0:0) to enable proper dependency resolution and building of the application.
Feedback submitted
