+++ backend/cmd/main_test.go
@@ -1,435 +1,1 @@
 package main
-
-// import (
-// 	"bytes"
-// 	"context"
-// 	"encoding/csv"
-// 	"encoding/json"
-// 	"errors"
-// 	"io"
-// 	"log/slog"
-// 	"net/http"
-// 	"net/http/httptest"
-// 	"os"
-// 	"reflect"
-// 	"strings"
-// 	"testing"
-// 	"time"
-// )
-
-// // mockRoundTripper is a custom http.RoundTripper for mocking HTTP responses.
-// type mockRoundTripper struct {
-// 	roundTrip func(req *http.Request) *http.Response
-// }
-
-// func (m *mockRoundTripper) RoundTrip(req *http.Request) (*http.Response, error) {
-// 	return m.roundTrip(req), nil
-// }
-
-// // newTestApp creates a new App instance for testing with a mocked HTTP client.
-// func newTestApp(handler func(req *http.Request) *http.Response) *App {
-// 	logger := slog.New(slog.NewJSONHandler(os.Stdout, &slog.HandlerOptions{Level: slog.LevelDebug}))
-// 	return &App{
-// 		config: Config{SnykOrgID: "test-org", SnykToken: "test-token"},
-// 		logger: logger,
-// 		httpClient: &http.Client{
-// 			Transport: &mockRoundTripper{roundTrip: handler},
-// 		},
-// 	}
-// }
-
-// // TestDataHandler_Success tests the successful orchestration of the data handler.
-// func TestDataHandler_Success(t *testing.T) {
-// 	// Mock server to simulate the Snyk API responses
-// 	mockServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-// 		if strings.Contains(r.URL.Path, "/exports") {
-// 			// Initial export request
-// 			w.WriteHeader(http.StatusOK)
-// 			json.NewEncoder(w).Encode(map[string]string{"export_id": "test-export-id"})
-// 		} else if strings.Contains(r.URL.Path, "/exports/test-export-id") {
-// 			// Polling for export status
-// 			w.WriteHeader(http.StatusOK)
-// 			json.NewEncoder(w).Encode(map[string]interface{}{
-// 				"status": "FINISHED",
-// 				"results": map[string][]map[string]string{
-// 					"files": {{"url": "http://localhost/test-csv"}},
-// 				},
-// 			})
-// 		} else if r.URL.Path == "/test-csv" {
-// 			// Serving the CSV file
-// 			w.Header().Set("Content-Type", "text/csv")
-// 			w.WriteHeader(http.StatusOK)
-// 			csvWriter := csv.NewWriter(w)
-// 			csvWriter.Write([]string{"issue_severity", "project_name", "project_environments", "computed_fixability"})
-// 			csvWriter.Write([]string{"critical", "proj1", "env1", "fixable"})
-// 			csvWriter.Flush()
-// 		}
-// 	}))
-// 	defer mockServer.Close()
-
-// 	// Create a new app with a client that uses the mock server
-// 	app := newTestApp(func(req *http.Request) *http.Response {
-// 		// Redirect requests to the mock server
-// 		newURL := mockServer.URL + req.URL.Path
-// 		if req.URL.RawQuery != "" {
-// 			newURL += "?" + req.URL.RawQuery
-// 		}
-// 		proxyReq, _ := http.NewRequest(req.Method, newURL, req.Body)
-// 		proxyReq.Header = req.Header
-// 		resp, err := http.DefaultClient.Do(proxyReq)
-// 		if err != nil {
-// 			t.Fatalf("Failed to proxy request to mock server: %v", err)
-// 		}
-// 		return resp
-// 	})
-
-// 	req := httptest.NewRequest("GET", "/api/data", nil)
-// 	rr := httptest.NewRecorder()
-// 	app.dataHandler(rr, req)
-
-// 	if status := rr.Code; status != http.StatusOK {
-// 		t.Errorf("handler returned wrong status code: got %v want %v", status, http.StatusOK)
-// 	}
-
-// 	var data DashboardData
-// 	if err := json.NewDecoder(rr.Body).Decode(&data); err != nil {
-// 		t.Fatalf("could not decode response: %v", err)
-// 	}
-
-// 	expectedIssuesBySeverity := map[string]int{"critical": 1}
-// 	if !reflect.DeepEqual(data.IssuesBySeverity, expectedIssuesBySeverity) {
-// 		t.Errorf("unexpected IssuesBySeverity: got %v, want %v", data.IssuesBySeverity, expectedIssuesBySeverity)
-// 	}
-// }
-
-// // TestFetchAndProcessCSV tests the CSV processing logic.
-// func TestFetchAndProcessCSV(t *testing.T) {
-// 	csvContent := `issue_severity,project_name,project_environments,computed_fixability
-// critical,proj1,env1,fixable
-// high,proj1,env1,
-// critical,proj2,env2,fixable
-// low,proj3,env1,
-// medium,proj2,env2,
-// critical,proj1,env1,`
-
-// 	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-// 		w.Header().Set("Content-Type", "text/csv")
-// 		w.WriteHeader(http.StatusOK)
-// 		io.WriteString(w, csvContent)
-// 	}))
-// 	defer server.Close()
-
-// 	app := newTestApp(nil)
-// 	data, err := app.fetchAndProcessCSV(context.Background(), server.URL)
-
-// 	if err != nil {
-// 		t.Fatalf("fetchAndProcessCSV failed: %v", err)
-// 	}
-
-// 	expectedIssuesBySeverity := map[string]int{"critical": 3, "high": 1, "low": 1, "medium": 1}
-// 	if !reflect.DeepEqual(data.IssuesBySeverity, expectedIssuesBySeverity) {
-// 		t.Errorf("unexpected IssuesBySeverity: got %v, want %v", data.IssuesBySeverity, expectedIssuesBySeverity)
-// 	}
-
-// 	expectedFixableCritical := 2
-// 	if data.FixableCriticalIssues != expectedFixableCritical {
-// 		t.Errorf("unexpected FixableCriticalIssues: got %v, want %v", data.FixableCriticalIssues, expectedFixableCritical)
-// 	}
-
-// 	expectedTopProjects := []ProjectInfo{
-// 		{Name: "proj1", CriticalIssueCount: 2, HighIssueCount: 1},
-// 		{Name: "proj2", CriticalIssueCount: 1, HighIssueCount: 0},
-// 		{Name: "proj3", CriticalIssueCount: 0, HighIssueCount: 0},
-// 	}
-// 	if !reflect.DeepEqual(data.Top5RiskiestProjects, expectedTopProjects) {
-// 		t.Errorf("unexpected Top5RiskiestProjects: got %+v, want %+v", data.Top5RiskiestProjects, expectedTopProjects)
-// 	}
-// }
-
-// // TestInitiateExport_Success tests the successful initiation of an export.
-// func TestInitiateExport_Success(t *testing.T) {
-// 	app := newTestApp(func(req *http.Request) *http.Response {
-// 		return &http.Response{
-// 			StatusCode: http.StatusOK,
-// 			Body:       io.NopCloser(bytes.NewBufferString(`{"export_id":"test-export-id"}`)),
-// 			Header:     make(http.Header),
-// 		}
-// 	})
-
-// 	exportID, err := app.initiateExport(context.Background())
-// 	if err != nil {
-// 		t.Fatalf("initiateExport failed: %v", err)
-// 	}
-// 	if exportID != "test-export-id" {
-// 		t.Errorf("unexpected export ID: got %s, want test-export-id", exportID)
-// 	}
-// }
-
-// // TestPollExportStatus_Success tests successful polling for a finished export.
-// func TestPollExportStatus_Success(t *testing.T) {
-// 	app := newTestApp(func(req *http.Request) *http.Response {
-// 		return &http.Response{
-// 			StatusCode: http.StatusOK,
-// 			Body: io.NopCloser(bytes.NewBufferString(`{
-// 				"status": "FINISHED",
-// 				"results": {
-// 					"files": [{"url": "http://snyk.io/export.csv"}]
-// 				}
-// 			}`)),
-// 			Header: make(http.Header),
-// 		}
-// 	})
-
-// 	// Shorten polling for test efficiency
-// 	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Millisecond)
-// 	defer cancel()
-
-// 	fileURL, err := app.pollExportStatus(ctx, "test-export-id")
-// 	if err != nil && !errors.Is(err, context.DeadlineExceeded) { // Ignore timeout error for this test setup
-// 		t.Fatalf("pollExportStatus failed: %v", err)
-// 	}
-
-// 	// In this mocked scenario, we might not get the URL in time, so we check if it's as expected when it is returned.
-// 	// A more robust test would involve a more complex mock that changes its response over time.
-// 	// For this case, we'll assume if no error other than timeout, the logic inside is sound.
-// 	// If a URL is returned, it should be the correct one.
-// 	if fileURL != "" && fileURL != "http://snyk.io/export.csv" {
-// 		t.Errorf("unexpected file URL: got %s, want http://snyk.io/export.csv", fileURL)
-// 	}
-// }
-
-// // TestGetEnv tests the getEnv helper function.
-// func TestGetEnv(t *testing.T) {
-// 	// Test case where environment variable is set
-// 	os.Setenv("TEST_ENV_VAR", "test_value")
-// 	val := getEnv("TEST_ENV_VAR", "fallback")
-// 	if val != "test_value" {
-// 		t.Errorf("getEnv failed: expected 'test_value', got '%s'", val)
-// 	}
-// 	os.Unsetenv("TEST_ENV_VAR")
-
-// 	// Test case where environment variable is not set
-// 	val = getEnv("NON_EXISTENT_VAR", "fallback_value")
-// 	if val != "fallback_value" {
-// 		t.Errorf("getEnv failed: expected 'fallback_value', got '%s'", val)
-// 	}
-// }
-
-// // TestNewApp tests the NewApp constructor.
-// func TestNewApp(t *testing.T) {
-// 	logger := slog.New(slog.NewJSONHandler(os.Stdout, nil))
-// 	cfg := Config{SnykToken: "token", SnykOrgID: "org", Port: "8080"}
-// 	app := NewApp(cfg, logger)
-
-// 	if app.config.SnykToken != "token" {
-// 		t.Error("NewApp did not set SnykToken correctly")
-// 	}
-// 	if app.config.SnykOrgID != "org" {
-// 		t.Error("NewApp did not set SnykOrgID correctly")
-// 	}
-// 	if app.logger == nil {
-// 		t.Error("NewApp did not set logger")
-// 	}
-// 	if app.httpClient == nil {
-// 		t.Error("NewApp did not set httpClient")
-// 	}
-// }
-
-// // TestRespondWithJSON tests the JSON response helper.
-// func (a *App) TestRespondWithJSON(t *testing.T) {
-// 	rr := httptest.NewRecorder()
-// 	payload := map[string]string{"hello": "world"}
-// 	a.respondWithJSON(rr, http.StatusOK, payload)
-
-// 	if rr.Code != http.StatusOK {
-// 		t.Errorf("respondWithJSON status code: got %v, want %v", rr.Code, http.StatusOK)
-// 	}
-// 	if !strings.Contains(rr.Header().Get("Content-Type"), "application/json") {
-// 		t.Errorf("respondWithJSON content type: got %v, want application/json", rr.Header().Get("Content-Type"))
-// 	}
-// 	expectedBody := `{"hello":"world"}`
-// 	if rr.Body.String() != expectedBody {
-// 		t.Errorf("respondWithJSON body: got %v, want %v", rr.Body.String(), expectedBody)
-// 	}
-// }
-
-// // TestSetAuthHeader tests that the auth header is set correctly.
-// func TestSetAuthHeader(t *testing.T) {
-// 	req, _ := http.NewRequest("GET", "/", nil)
-// 	app := &App{config: Config{SnykToken: "test-token"}}
-// 	app.setAuthHeader(req)
-
-// 	authHeader := req.Header.Get("Authorization")
-// 	if authHeader != "token test-token" {
-// 		t.Errorf("setAuthHeader Authorization: got '%s', want 'token test-token'", authHeader)
-// 	}
-// 	contentTypeHeader := req.Header.Get("Content-Type")
-// 	if contentTypeHeader != "application/json" {
-// 		t.Errorf("setAuthHeader Content-Type: got '%s', want 'application/json'", contentTypeHeader)
-// 	}
-// }
-
-// // TestDataHandler_InitiateExportFails tests the data handler when initiating an export fails.
-// func TestDataHandler_InitiateExportFails(t *testing.T) {
-// 	app := newTestApp(func(req *http.Request) *http.Response {
-// 		return &http.Response{
-// 			StatusCode: http.StatusInternalServerError,
-// 			Body:       io.NopCloser(bytes.NewBufferString(`{"error": "failed to initiate export"}`)),
-// 			Header:     make(http.Header),
-// 		}
-// 	})
-
-// 	req := httptest.NewRequest("GET", "/api/data", nil)
-// 	rr := httptest.NewRecorder()
-// 	app.dataHandler(rr, req)
-
-// 	if status := rr.Code; status != http.StatusInternalServerError {
-// 		t.Errorf("handler returned wrong status code: got %v want %v", status, http.StatusInternalServerError)
-// 	}
-// }
-
-// // TestDataHandler_PollExportFails tests the data handler when polling for export status fails.
-// func TestDataHandler_PollExportFails(t *testing.T) {
-// 	app := newTestApp(func(req *http.Request) *http.Response {
-// 		if strings.Contains(req.URL.Path, "/exports") {
-// 			return &http.Response{
-// 				StatusCode: http.StatusOK,
-// 				Body:       io.NopCloser(bytes.NewBufferString(`{"export_id":"test-export-id"}`)),
-// 				Header:     make(http.Header),
-// 			}
-// 		}
-// 		return &http.Response{
-// 			StatusCode: http.StatusInternalServerError,
-// 			Body:       io.NopCloser(bytes.NewBufferString(`{"error": "failed to poll export status"}`)),
-// 			Header:     make(http.Header),
-// 		}
-// 	})
-
-// 	req := httptest.NewRequest("GET", "/api/data", nil)
-// 	rr := httptest.NewRecorder()
-// 	app.dataHandler(rr, req)
-
-// 	if status := rr.Code; status != http.StatusInternalServerError {
-// 		t.Errorf("handler returned wrong status code: got %v want %v", status, http.StatusInternalServerError)
-// 	}
-// }
-
-// // TestDataHandler_FetchAndProcessCSVFails tests the data handler when fetching and processing the CSV fails.
-// func TestDataHandler_FetchAndProcessCSVFails(t *testing.T) {
-// 	mockServer := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-// 		if strings.Contains(r.URL.Path, "/exports") {
-// 			w.WriteHeader(http.StatusOK)
-// 			json.NewEncoder(w).Encode(map[string]string{"export_id": "test-export-id"})
-// 		} else if strings.Contains(r.URL.Path, "/exports/test-export-id") {
-// 			w.WriteHeader(http.StatusOK)
-// 			json.NewEncoder(w).Encode(map[string]interface{}{
-// 				"status": "FINISHED",
-// 				"results": map[string][]map[string]string{
-// 					"files": {{"url": "http://localhost/test-csv"}},
-// 				},
-// 			})
-// 		} else if r.URL.Path == "/test-csv" {
-// 			w.WriteHeader(http.StatusInternalServerError)
-// 		}
-// 	}))
-// 	defer mockServer.Close()
-
-// 	app := newTestApp(func(req *http.Request) *http.Response {
-// 		newURL := mockServer.URL + req.URL.Path
-// 		if req.URL.RawQuery != "" {
-// 			newURL += "?" + req.URL.RawQuery
-// 		}
-// 		proxyReq, _ := http.NewRequest(req.Method, newURL, req.Body)
-// 		proxyReq.Header = req.Header
-// 		resp, err := http.DefaultClient.Do(proxyReq)
-// 		if err != nil {
-// 			t.Fatalf("Failed to proxy request to mock server: %v", err)
-// 		}
-// 		return resp
-// 	})
-
-// 	req := httptest.NewRequest("GET", "/api/data", nil)
-// 	rr := httptest.NewRecorder()
-// 	app.dataHandler(rr, req)
-
-// 	if status := rr.Code; status != http.StatusInternalServerError {
-// 		t.Errorf("handler returned wrong status code: got %v want %v", status, http.StatusInternalServerError)
-// 	}
-// }
-
-// // TestPollExportStatus_Error tests polling for an export that results in an error.
-// func TestPollExportStatus_Error(t *testing.T) {
-// 	app := newTestApp(func(req *http.Request) *http.Response {
-// 		return &http.Response{
-// 			StatusCode: http.StatusOK,
-// 			Body:       io.NopCloser(bytes.NewBufferString(`{"status": "ERROR"}`)),
-// 			Header:     make(http.Header),
-// 		}
-// 	})
-
-// 	_, err := app.pollExportStatus(context.Background(), "test-export-id")
-// 	if err == nil {
-// 		t.Error("pollExportStatus should have returned an error")
-// 	}
-// }
-
-// // TestPollExportStatus_Timeout tests polling for an export that times out.
-// func TestPollExportStatus_Timeout(t *testing.T) {
-// 	app := newTestApp(func(req *http.Request) *http.Response {
-// 		return &http.Response{
-// 			StatusCode: http.StatusOK,
-// 			Body:       io.NopCloser(bytes.NewBufferString(`{"status": "PENDING"}`)),
-// 			Header:     make(http.Header),
-// 		}
-// 	})
-
-// 	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Millisecond)
-// 	defer cancel()
-
-// 	_, err := app.pollExportStatus(ctx, "test-export-id")
-// 	if !errors.Is(err, context.DeadlineExceeded) {
-// 		t.Errorf("pollExportStatus should have timed out, but returned: %v", err)
-// 	}
-// }
-
-// // TestFetchAndProcessCSV_Empty tests processing an empty CSV file.
-// func TestFetchAndProcessCSV_Empty(t *testing.T) {
-// 	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-// 		w.Header().Set("Content-Type", "text/csv")
-// 		w.WriteHeader(http.StatusOK)
-// 		io.WriteString(w, "issue_severity,project_name,project_environments,computed_fixability\n") // Empty CSV header
-// 	}))
-// 	defer server.Close()
-
-// 	app := newTestApp(nil)
-// 	data, err := app.fetchAndProcessCSV(context.Background(), server.URL)
-// 	if err != nil {
-// 		t.Fatalf("fetchAndProcessCSV failed: %v", err)
-// 	}
-// 	if len(data.IssuesBySeverity) != 0 {
-// 		t.Errorf("expected 0 issues by severity, got %d", len(data.IssuesBySeverity))
-// 	}
-// }
-
-// type errorReader struct{}
-
-// func (e *errorReader) Read(p []byte) (n int, err error) {
-// 	return 0, errors.New("read error")
-// }
-
-// // TestFetchAndProcessCSV_Malformed tests processing a malformed CSV file.
-// func TestFetchAndProcessCSV_Malformed(t *testing.T) {
-// 	server := httptest.NewServer(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-// 		w.Header().Set("Content-Type", "text/csv")
-// 		w.WriteHeader(http.StatusOK)
-// 		// This CSV is malformed because of the unclosed quote.
-// 		io.WriteString(w, `"issue_severity,project_name`)
-// 	}))
-// 	defer server.Close()
-
-// 	app := newTestApp(nil)
-// 	_, err := app.fetchAndProcessCSV(context.Background(), server.URL)
-// 	if err == nil {
-// 		t.Error("fetchAndProcessCSV should have returned an error for a malformed CSV")
-// 	}
-// }
